% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{build_output}
\alias{build_output}
\title{Builds the output structure.}
\usage{
build_output(output, wait)
}
\arguments{
\item{output}{The invocation output}

\item{wait}{A TRUE/FALSE parameter, indicating whether the function should wait for the command to finish, or run it asynchronously}
}
\value{
The script output structure
}
\description{
Builds the output structure.
}
\examples{
output <- c('line 1', '\\n', 'line 2')
attr(output, 'status') <- 15
script_output <- build_output(output)
}
