% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{create_system_call_args}
\alias{create_system_call_args}
\title{Returns the system call arguments.}
\usage{
create_system_call_args(command, cli_args, wait, env, is_windows_os)
}
\arguments{
\item{command}{The command to invoke}

\item{cli_args}{Possible list of command line arguments}

\item{wait}{A TRUE/FALSE parameter, indicating whether the function should wait for the command to finish, or run it asynchronously}

\item{env}{Optional character vector of name=value strings to set environment variables}

\item{is_windows_os}{True if windows based OS, false for unix based OS}
}
\value{
The system call arguments
}
\description{
Returns the system call arguments.
}
\examples{
filename <- './myfile.sh'
arg_list <- create_system_call_args('sh', c(filename), TRUE, character(), FALSE)
}
