% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{create_temp_file}
\alias{create_temp_file}
\title{Creates a temporary file, writes the provided script content into it and returns the file name.}
\usage{
create_temp_file(script)
}
\arguments{
\item{script}{The script text}
}
\value{
The temporary file name
}
\description{
Creates a temporary file, writes the provided script content into it and returns the file name.
}
