% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{generate_args_setup_script}
\alias{generate_args_setup_script}
\title{Generates and returns a script which sets up the env vars for the script arguments}
\usage{
generate_args_setup_script(args = character())
}
\arguments{
\item{args}{Optional script command line arguments}
}
\value{
The script text which sets up the env vars for the script arguments
}
\description{
Generates and returns a script which sets up the env vars for the script arguments
}
\examples{
script <- generate_args_setup_script(args = c('first', 'second'))
}
