% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{generate_env_setup_script}
\alias{generate_env_setup_script}
\title{Generates and returns a script which sets up the env vars for the script execution.}
\usage{
generate_env_setup_script(env = character())
}
\arguments{
\item{env}{Optional character vector of name=value strings to set environment variables}
}
\value{
The script text which sets up the env
}
\description{
Generates and returns a script which sets up the env vars for the script execution.
}
\examples{
script <- generate_env_setup_script(c('ENV_TEST=MYENV'))
}
