% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{get_command}
\alias{get_command}
\title{Returns the command and arguments needed to execute the provided script file on the current platform.}
\usage{
get_command(filename)
}
\arguments{
\item{filename}{The script file to execute}
}
\value{
A list holding the command and arguments
}
\description{
Returns the command and arguments needed to execute the provided script file on the current platform.
}
