% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{is_windows}
\alias{is_windows}
\title{Returns true if windows, else false.}
\usage{
is_windows()
}
\value{
True if windows, else false.
}
\description{
Returns true if windows, else false.
}
\examples{
windows <- is_windows()
}
