% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{on_invoke_error}
\alias{on_invoke_error}
\title{Internal error handler.}
\usage{
on_invoke_error(error)
}
\arguments{
\item{error}{The invocation error}
}
\value{
The invocation output
}
\description{
Internal error handler.
}
