% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{script_execute}
\alias{script_execute}
\title{Executes a script and returns the output.
The stdout and stderr are captured and returned.
In case of errors, the exit code will return in the status field.}
\usage{
script_execute(script, args = c(), env = character())
}
\arguments{
\item{script}{The script text}

\item{args}{Optional script command line arguments (arguments are added as variables in the script named ARG1, ARG2, ...)}

\item{env}{Optional character vector of name=value strings to set environment variables}
}
\value{
The script output, see system2 documentation
}
\description{
Executes a script and returns the output.
The stdout and stderr are captured and returned.
In case of errors, the exit code will return in the status field.
}
\examples{
#execute script text
output <- script_execute('echo Current Directory:\\ndir')
cat(sprintf('Exit Status: \%s Output\%s\\n', output$status, output$output))

#execute multiple commands as a script
output <- script_execute(c('cd', 'echo User Home:', 'dir'))
cat(sprintf('Exit Status: \%s Output\%s\\n', output$status, output$output))

#pass argument to the script, later defined as ARG1
output <- script_execute(c('echo $ARG1 $ARG2'), c('TEST1', 'TEST2'))
cat(sprintf('\%s\\n', output))

#status is returned in case of errors
output <- script_execute('exit 1')
cat(sprintf('Status: \%s\\n', output$status))
cat(sprintf('\%s\\n', output))
}
