% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\docType{package}
\name{scriptexec}
\alias{scriptexec}
\alias{scriptexec-package}
\title{scriptexec: Execute native scripts}
\description{
This package provides one main function: script_execute which executes the provided script and returns its output.
}
