% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{get_command}
\alias{get_command}
\title{Returns the command and arguments needed to execute the provided script file on the current platform.}
\usage{
get_command(filename, runner = NULL)
}
\arguments{
\item{filename}{The script file to execute}

\item{runner}{The executable used to invoke the script (by default cmd.exe for windows, sh for other platforms)}
}
\value{
A list holding the command and arguments
}
\description{
Returns the command and arguments needed to execute the provided script file on the current platform.
}
\examples{
command_struct <- get_command('myfile.sh')
command <- command_struct$command
cli_args <- command_struct$args
}
