% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{modify_script}
\alias{modify_script}
\title{Modifies the provided script text and ensures the script content is executed in the correct location.}
\usage{
modify_script(script, args = c())
}
\arguments{
\item{script}{The script text}

\item{args}{Optional script command line arguments}
}
\value{
The modified script text
}
\description{
Modifies the provided script text and ensures the script content is executed in the correct location.
}
\examples{
full.script <- modify_script(script = script, args = args)
}
