% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{modify_script}
\alias{modify_script}
\title{Modifies the provided script text and ensures the script content is executed in the correct location.}
\usage{
modify_script(script)
}
\arguments{
\item{script}{The script text}
}
\value{
The modified script text
}
\description{
Modifies the provided script text and ensures the script content is executed in the correct location.
}
