% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptexec.R
\name{script_execute}
\alias{script_execute}
\title{Executes a script and returns the output.
The stdout and stderr are captured and returned.
In case of errors, the exit code will return in the status field.}
\usage{
script_execute(script, args = c())
}
\arguments{
\item{script}{The script text}

\item{args}{Optional script command line arguments}
}
\value{
The script output, see system2 documentation
}
\description{
Executes a script and returns the output.
The stdout and stderr are captured and returned.
In case of errors, the exit code will return in the status field.
}
\examples{
output <- script_execute('echo Current Directory:\\ndir') #execute script text
cat(sprintf('\%s\\n', output))
output <- script_execute(c('cd', 'echo User Home:', 'dir')) #execute multiple commands as a script
cat(sprintf('\%s\\n', output))
}
